+++ modules/client/src/controllers/index.ts
@@ -1,5 +1,4 @@
 export { DepositController } from "./DepositController";
 export { SwapController } from "./SwapController";
 export { WithdrawalController } from "./WithdrawalController";
-export { CreateTransferController } from "./CreateTransferController";
-export { ResolveTransferController } from "./ResolveTransferController";
+export { TransferController } from "./TransferController";
